ts2pant: translate empty ReadonlyMap fields; extract translate-record.ts
Two changes in this commit.

1) Dogfood unblock. isMapType now recognizes ReadonlyMap (the Set twin
   already handled both Set and ReadonlySet), so interface fields typed
   ReadonlyMap<K, V> hit the Stage A field-Map encoding instead of
   falling through to the generic-rule path with raw "ReadonlyMap" as
   the return type. Add a symmetric empty-Map branch to the record-
   return emitter: `new Map()` in an interface field emits
   `all k: K | ~(<field>Key receiver k)`, mirroring the existing empty-
   set `all x: T | ~(x in <field> receiver)`. collectNamedTypes now
   recurses into Set/Map type arguments too (previously only array /
   tuple), so the V-type gets declared as a domain. Expands dogfood
   coverage to emptyMapSynth + emptyRecordSynth.

2) Extract record-return emission into its own module. translateRecord-
   Return, emitRecordEquations, and the empty-Set/Map helpers move to
   src/translate-record.ts (~526 LOC). translate-body.ts drops from
   4331 → 3825 LOC and exports the minimal surface the new module
   needs: UniqueSupply, isBodyUnsupported, isBodyEffect, bodyExpr.
   Behavior preserved; 324 tests pass.
